# BackEnd/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Zurqa Al-Yamama API",
        "description": "API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://zurqa-al-yamama.onrender.com",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/api/reports/analyze": {
            "post": {
                "tags": [
                    "Reports"
                ],
                "summary": "Analyze an uploaded report (PDF or image)",
                "operationId": "07df7f3e8c9945b16a66ab7d8d85adde",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "report": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Analysis result"
                    }
                }
            }
        },
...
        "/api/texts/analyze": {
            "post": {
                "tags": [
                    "Texts"
                ],
                "summary": "Analyze plain text",
                "operationId": "a011f6a3324c874af2a5dd9ac8df3fec",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "text": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Analysis result"
                    }
                }
            }
        },
...
    },
...
}